# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /recipes:
    x-swagger-router-controller: recipes
    get:
      tags: 
      - recipe
      summary: Views recipes
      description: Get all the recipes
      operationId: getRecipe
      responses:
        "200":
          description: Success, got all the recipes
          schema:
            type: array
            items: 
              $ref: "#/definitions/Recipe"
        "500":
          description: Could not get recipes
          schema:
            type: object
            properties:
              message:
                type: string
    post:
      tags: 
      - recipe
      summary: Create new recipe
      operationId: postRecipe
      parameters:
        - in: body
          name: body
          description: The recipe to be added
          required: true
          schema:
            $ref: "#/definitions/Recipe"
      responses: 
        "204":
          description: Success
        "500":
          description: Unexpected Error
          schema:
            type: object
            properties:
              message:
                type: string
...
definitions:
...
  Recipe:
    properties:
      id:
        type: number
      isPublic:
        type: boolean
      name: 
        type: string
      chef:
        type: string
        enum: [cook, culinarian]
      stepList:
        type: array
        items: 
          type: string
      ingredients:
          type: array
          items: 
            type: object
            properties:
              amount:
                type: number
              name: 
                type: string
...